# Speed up loading page.