--- a/src/controllers/root_controller.js
+++ b/src/controllers/root_controller.js
@@ -21,7 +21,7 @@ const RootController = express.Router();
  */
 RootController.get('/', TunnelSetup.isTunnelSet, function(request, response) {
   response.sendFile('index.html', {
-    root: Constants.VIEWS_PATH,
+    root: Constants.BUILD_STATIC_PATH,
   });
 });