--- a/src/views/authorize.mustache
+++ b/src/views/authorize.mustache
@@ -14,8 +14,6 @@
     <link rel="stylesheet" type="text/css" href="/css/thing.css" />
     <link rel="stylesheet" type="text/css" href="/css/authorize.css" />
 
-    <script type="text/javascript" src="/js/utils.js"></script>
-    <script type="text/javascript" src="/js/api.js"></script>
     <script type="text/javascript" src="/js/check-user.js"></script>
     <script type="text/javascript" src="/js/authorize.js"></script>
   </head>