--- a/src/views/local-token-service.mustache
+++ b/src/views/local-token-service.mustache
@@ -17,8 +17,6 @@
     <link rel="stylesheet" type="text/css" href="/css/local-token-service.css" />
 
     <script type="text/javascript" src="/js/lib/highlight-9.12.0.min.js"></script>
-    <script type="text/javascript" src="/js/utils.js"></script>
-    <script type="text/javascript" src="/js/api.js"></script>
     <script type="text/javascript" src="/js/check-user.js"></script>
   </head>
   <body class="hidden">