--- a/static/js/addons-config/error-field.js
+++ b/static/js/addons-config/error-field.js
@@ -48,3 +48,5 @@ ErrorField.prototype.render = function(errors) {
 
   return this.field;
 };
+
+module.exports = ErrorField;