In order to speed up loading web pages, i make the following changes.
* packing static files into one file.
* lazy loading some lib/*js files.

In my case the loading speed doubled.
I hope that this pull request will be effective in other's environment as well.