@mrstegeman 2018-03-29
@sogaani You'll also need to run `yarn` and commit the updated `yarn.lock` file.