@hobinjk 2018-03-29
I looked into this a while ago and the primary concern at the time was to ensure that the compiled files would have corresponding source maps so debugging isn't negatively effected. I'll take a look at this when it's in a more finished state.