@sogaani 2018-03-30
I figured out cause of fail.
We need to exec webpack before test.