@mrstegeman 2018-04-02
@sogaani I just rebased this patch set, since I just modified a ton of stuff with eslint. Hopefully I didn't break anything too badly.