@sogaani 2018-04-03
@mrstegeman  Thank you! There was no problem with rebase. I fixed my part because there was a problem. (module.exports of Rule Engin)