@sogaani 2018-04-16
rebase to resolve conflict.
@hobinjk Could you check this pull request? I believe that this is already work.