@hobinjk 2018-04-17
This looks good to me, everything works as it's expected to except for the recently added src/view/local-token-service.mustache (e.g. https://localhost:4443/oauth/local-token-service?code=helloworld).

While I do have some concerns about the extra steps this adds to the build process, I think that the preservation of debugging information is really great. The performance characteristics of this are also very impressive, reducing the time-til-interactive by 16s to 6s on first view and 11s to 4s on repeat.

<img width="871" alt="image of console showing debug information" src="https://user-images.githubusercontent.com/2776089/38896722-8d769c7c-4261-11e8-9186-46d207af4cff.png">

before: https://www.webpagetest.org/result/180417_CD_1af6488672afee15c840d8b87e12c17b/
after: https://www.webpagetest.org/result/180417_XQ_efe5c779f5acec1ee75f93581e71215d/