@sogaani 2018-04-18
Fix conflict and typo.

@hobinjk Thank you for reviewing.
> to except for the recently added src/view/local-token-service.mustache
Maybe we need to access ```https://localhost:4443/oauth/authorize?response_type=code&client_id=local-token&scope=/things:readwrite&state=asdf```.
Above URL works for me.